refactor: disk manager sink, stream and builder
